Update get_outstanding_fines_with_user.sql
numeric

+++ queries/get_outstanding_fines_with_user.sql
@@ -6,8 +6,8 @@ RETURNS TABLE
 (
     action_date DATE,
     action_type TEXT,    
-    bill_amount INTEGER,
-    account_balance INTEGER,
+    bill_amount NUMERIC,
+    account_balance NUMERIC,
     user_barcode TEXT,
     username TEXT,
     user_last_name TEXT,
